Refactor: 필드명 및 스타일 경로 수정
